# libs/API/config/routes.php
<?php

declare(strict_types=1);

use Psr\Http\Message\ResponseFactoryInterface;
use Yiisoft\Csrf\CsrfMiddleware;
use Yiisoft\DataResponse\Middleware\FormatDataResponseAsJson;
use Yiisoft\Router\Group;
use Yiisoft\Router\Route;
use Yiisoft\Validator\ValidatorInterface;
use AppDevPanel\Api\Debug\Controller\DebugController;
use AppDevPanel\Api\Debug\Middleware\ResponseDataWrapper;
use AppDevPanel\Api\Debug\Middleware\TokenAuthMiddleware;
use AppDevPanel\Api\Ingestion\Controller\IngestionController;
use AppDevPanel\Api\Inspector\Controller\CacheController;
use AppDevPanel\Api\Inspector\Controller\ServiceController;
use AppDevPanel\Api\Inspector\Middleware\InspectorProxyMiddleware;
use AppDevPanel\Api\Inspector\Controller\CommandController;
use AppDevPanel\Api\Inspector\Controller\ComposerController;
use AppDevPanel\Api\Inspector\Controller\DatabaseController;
use AppDevPanel\Api\Inspector\Controller\FileController;
use AppDevPanel\Api\Inspector\Controller\GitController;
use AppDevPanel\Api\Inspector\Controller\InspectController;
use AppDevPanel\Api\Inspector\Controller\OpcacheController;
use AppDevPanel\Api\Inspector\Controller\RequestController;
use AppDevPanel\Api\Inspector\Controller\RoutingController;
use AppDevPanel\Api\Inspector\Controller\TranslationController;
use Yiisoft\Yii\Middleware\CorsAllowAll;
use Yiisoft\Yii\Middleware\IpFilter;

if (!(bool) ($params['app-dev-panel/yii-debug-api']['enabled'] ?? false)) {
    return [];
}

return [
    Group::create('/debug/api')
        ->withCors(CorsAllowAll::class)
        ->disableMiddleware(CsrfMiddleware::class)
        ->middleware(
            static function (ResponseFactoryInterface $responseFactory, ValidatorInterface $validator) use ($params) {
                return new IpFilter(
                    validator: $validator,
                    responseFactory: $responseFactory,
                    ipRanges: $params['app-dev-panel/yii-debug-api']['allowedIPs']
                );
            }
        )
        ->middleware(TokenAuthMiddleware::class)
        ->middleware(FormatDataResponseAsJson::class)
        ->middleware(ResponseDataWrapper::class)
        ->namePrefix('debug/api/')
        ->routes(
            Route::get('[/]')
                ->action([DebugController::class, 'index'])
                ->name('index'),
            Route::get('/summary/{id}')
                ->action([DebugController::class, 'summary'])
                ->name('summary'),
            Route::get('/view/{id}')
                ->action([DebugController::class, 'view'])
                ->name('view'),
            Route::get('/dump/{id}')
                ->action([DebugController::class, 'dump'])
                ->name('dump'),
            Route::get('/object/{id}/{objectId}')
                ->action([DebugController::class, 'object'])
                ->name('object'),
            Route::get('/event-stream')
                ->action([DebugController::class, 'eventStream'])
                ->name('event-stream'),
            Route::post('/ingest')
                ->action([IngestionController::class, 'ingest'])
                ->name('ingest'),
            Route::post('/ingest/batch')
                ->action([IngestionController::class, 'ingestBatch'])
                ->name('ingest/batch'),
            Route::post('/ingest/log')
                ->action([IngestionController::class, 'ingestLog'])
                ->name('ingest/log'),
            Route::get('/openapi.json')
                ->action([IngestionController::class, 'openapi'])
                ->name('openapi'),
            Group::create('/services')
                ->namePrefix('services/')
                ->routes(
                    Route::post('/register')
                        ->action([ServiceController::class, 'register'])
                        ->name('register'),
                    Route::post('/heartbeat')
                        ->action([ServiceController::class, 'heartbeat'])
                        ->name('heartbeat'),
                    Route::get('[/]')
                        ->action([ServiceController::class, 'list'])
                        ->name('list'),
                    Route::delete('/{service}')
                        ->action([ServiceController::class, 'deregister'])
                        ->name('deregister'),
                ),
        ),
    Group::create('/inspect/api')
        ->withCors(CorsAllowAll::class)
        ->disableMiddleware(CsrfMiddleware::class)
        ->middleware(
            static function (ResponseFactoryInterface $responseFactory, ValidatorInterface $validator) use ($params) {
                return new IpFilter(
                    validator: $validator,
                    responseFactory: $responseFactory,
                    ipRanges: $params['app-dev-panel/yii-debug-api']['allowedIPs']
                );
            }
        )
        ->middleware(TokenAuthMiddleware::class)
        ->middleware(FormatDataResponseAsJson::class)
        ->middleware(ResponseDataWrapper::class)
        ->middleware(InspectorProxyMiddleware::class)
        ->namePrefix('inspect/api/')
        ->routes(
            Route::get('/events')
                ->action([InspectController::class, 'eventListeners'])
                ->name('events'),
            Route::get('/params')
                ->action([InspectController::class, 'params'])
                ->name('params'),
            Route::get('/config')
                ->action([InspectController::class, 'config'])
                ->name('config'),
            Route::get('/classes')
                ->action([InspectController::class, 'classes'])
                ->name('classes'),
            Route::get('/object')
                ->action([InspectController::class, 'object'])
                ->name('object'),
            Route::get('/files')
                ->action([FileController::class, 'files'])
                ->name('files'),
            Route::get('/routes')
                ->action([RoutingController::class, 'routes'])
                ->name('routes'),
            Route::get('/route/check')
                ->action([RoutingController::class, 'checkRoute'])
                ->name('route/check'),
            Route::get('/translations')
                ->action([TranslationController::class, 'getTranslations'])
                ->name('getTranslations'),
            Route::put('/translations')
                ->action([TranslationController::class, 'putTranslation'])
                ->name('putTranslation'),
            Route::get('/table')
                ->action([DatabaseController::class, 'getTables'])
                ->name('getTables'),
            Route::get('/table/{name}')
                ->action([DatabaseController::class, 'getTable'])
                ->name('getTable'),
            Route::put('/request')
                ->action([RequestController::class, 'request'])
                ->name('request'),
            Route::post('/curl/build')
                ->action([RequestController::class, 'buildCurl'])
                ->name('curl/build'),
            Group::create('/git')
                ->namePrefix('/git')
                ->routes(
                    Route::get('/summary')
                        ->action([GitController::class, 'summary'])
                        ->name('summary'),
                    Route::post('/checkout')
                        ->action([GitController::class, 'checkout'])
                        ->name('checkout'),
                    Route::post('/command')
                        ->action([GitController::class, 'command'])
                        ->name('command'),
                    Route::get('/log')
                        ->action([GitController::class, 'log'])
                        ->name('log'),
                ),
            Route::get('/phpinfo')
                ->action([InspectController::class, 'phpinfo'])
                ->name('phpinfo'),
            Group::create('/command')
                ->namePrefix('command')
                ->routes(
                    Route::get('[/]')
                        ->action([CommandController::class, 'index'])
                        ->name('/index'),
                    Route::post('[/]')
                        ->action([CommandController::class, 'run'])
                        ->name('/run'),
                ),
            Group::create('/composer')
                ->namePrefix('composer')
                ->routes(
                    Route::get('[/]')
                        ->action([ComposerController::class, 'index'])
                        ->name('/index'),
                    Route::get('/inspect')
                        ->action([ComposerController::class, 'inspect'])
                        ->name('/inspect'),
                    Route::post('/require')
                        ->action([ComposerController::class, 'require'])
                        ->name('/require'),
                ),
            Group::create('/cache')
                ->namePrefix('cache')
                ->routes(
                    Route::get('[/]')
                        ->action([CacheController::class, 'view'])
                        ->name('/view'),
                    Route::delete('[/]')
                        ->action([CacheController::class, 'delete'])
                        ->name('/delete'),
                    Route::post('/clear')
                        ->action([CacheController::class, 'clear'])
                        ->name('/clear'),
                ),
            Group::create('/opcache')
                ->namePrefix('opcache')
                ->routes(
                    Route::get('[/]')
                        ->action([OpcacheController::class, 'index'])
                        ->name('/index'),
                ),
        ),
];